Feature/part6

+++ app/controllers/questions_controller.rb
@@ -1,10 +1,13 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
 
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer = question.answers.new
+  end
 
   def new; end
 
@@ -12,9 +15,10 @@ class QuestionsController < ApplicationController
 
   def create
     @question = Question.new(question_params)
+    @question.user = current_user
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -29,7 +33,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    question.destroy if current_user.author?(question)
     redirect_to questions_path
   end
